Mapped Indoor Navigation to PopUp Modals
