+++ ConcordiaMaps/components/MapMarkers.js
@@ -3,14 +3,21 @@ import { Marker, Callout } from "react-native-maps";
 import { Text, View, Image, StyleSheet } from "react-native";
 import PropTypes from "prop-types";
 import PopupModal from "./PopupModal";
-
+import { useNavigation, useRoute } from "@react-navigation/native";
+import styles from "../styles";
+import { Building } from "../constants/Building";
 const customMarkerImage = require("../assets/PinLogo.png");
 
-const MapMarkers = ({ markers }) => {
+const MapMarkers = () => {
   const [popupVisible, setPopupVisible] = useState(false);
   const [popupData, setPopupData] = useState(null);
-
-  if (!markers || markers.length === 0) return null;
+  const navigation = useNavigation();
+  
+  // if (!markers || markers.length === 0) 
+  //   {alert("Bombaclat");
+  //     return null;
+      
+  //   }
 
   const handleMarkerPress = (marker) => {
     setPopupData(marker);
@@ -24,7 +31,7 @@ const MapMarkers = ({ markers }) => {
 
   return (
     <>
-      {markers.map((marker) => (
+      {Building.map((marker) => (
         <Marker
           key={`${marker.name}-${marker.coordinate.latitude}-${marker.coordinate.longitude}`}
           coordinate={marker.coordinate}
@@ -44,699 +51,28 @@ const MapMarkers = ({ markers }) => {
         </Marker>
       ))}
 
-      <PopupModal
+      {/* <PopupModal
         isVisible={popupVisible}
         data={popupData}
         onClose={closePopup}
-      />
+        navigation = {navigation}
+      /> */}
     </>
   );
 };
 
-const styles = StyleSheet.create({
-  markerImage: {
-    width: 40,
-    height: 40,
-  },
-  calloutContainer: {
-    width: 160,
-    height: 50,
-    padding: 5,
-    backgroundColor: "rgba(255, 255, 255, 0.9)",
-    borderRadius: 8,
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  calloutText: {
-    fontSize: 14,
-    color: "#333",
-    textAlign: "center",
-  },
-});
 
-MapMarkers.propTypes = {
-  markers: PropTypes.arrayOf(
-    PropTypes.shape({
-      coordinate: PropTypes.shape({
-        latitude: PropTypes.number.isRequired,
-        longitude: PropTypes.number.isRequired,
-      }).isRequired,
-      name: PropTypes.string.isRequired,
-    }),
-  ),
-};
+// MapMarkers.propTypes = {
+//   markers: PropTypes.arrayOf(
+//     PropTypes.shape({
+//       coordinate: PropTypes.shape({
+//         latitude: PropTypes.number.isRequired,
+//         longitude: PropTypes.number.isRequired,
+//       }).isRequired,
+//       name: PropTypes.string.isRequired,
+//     }),
+//   ),
+// };
 
 export default MapMarkers;
 
-export const Building = [
-  // SGW Campus buildings latitude and longitude
-  {
-    name: "Webster Library",
-    coordinate: {
-      latitude: 45.4968158,
-      longitude: -73.5779337,
-    },
-    address: "1400 De Maisonneuve Blvd W, Montreal, QC H3G 1M8",
-    fullBuildingName: "Webster Library",
-  },
-  {
-    name: "B Annex",
-    coordinate: {
-      latitude: 45.497856,
-      longitude: -73.579588,
-    },
-    address: "2160 Bishop Street, Montreal, QC H3G 2E9",
-    fullBuildingName: "Bishop Annex",
-  },
-  {
-    name: "CL Annex",
-    coordinate: {
-      latitude: 45.494259,
-      longitude: -73.579007,
-    },
-    address: "1665 Ste-Catherine W, Montreal, QC H3H 1L7",
-    fullBuildingName: "Catherine-Lamb Annex",
-  },
-  {
-    name: "CI Annex",
-    coordinate: {
-      latitude: 45.497467,
-      longitude: -73.579925,
-    },
-    address: "2149 Mackay Street, Montreal, QC H3G 2E2",
-    fullBuildingName: "Concordia International Annex",
-  },
-  {
-    name: "D Annex",
-    coordinate: {
-      latitude: 45.497827,
-      longitude: -73.579409,
-    },
-    address: "2140 Bishop Street, Montreal, QC H3G 2E8",
-    fullBuildingName: "Duncan Annex",
-  },
-  {
-    name: "EN Annex",
-    coordinate: {
-      latitude: 45.496944,
-      longitude: -73.579555,
-    },
-    address: "2070 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Engineering Annex",
-  },
-  {
-    name: "ER Building",
-    coordinate: {
-      latitude: 45.496428,
-      longitude: -73.580212,
-    },
-    address: "2155 Guy Street, Montreal, QC H3H 2L9",
-    fullBuildingName: "Engineering Research Building",
-  },
-  {
-    name: "ES Building",
-    coordinate: {
-      latitude: 45.496172,
-      longitude: -73.579922,
-    },
-    address: "2135 Guy Street, Montreal, QC H3H 2L8",
-    fullBuildingName: "Engineering Science Building",
-  },
-  {
-    name: "ET Building",
-    coordinate: {
-      latitude: 45.496163,
-      longitude: -73.579904,
-    },
-    address: "2125-2127 Guy Street, Montreal, QC H3H 2L7",
-    fullBuildingName: "Engineering Technology Building",
-  },
-  {
-    name: "EV Building",
-    coordinate: {
-      latitude: 45.495376,
-      longitude: -73.577997,
-    },
-    address: "1515 Ste-Catherine W, Montreal, QC H3G 2W1",
-    fullBuildingName: "Engineering and Visual Arts Building",
-  },
-  {
-    name: "FA Annex",
-    coordinate: {
-      latitude: 45.496874,
-      longitude: -73.579468,
-    },
-    address: "2060 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Fine Arts Annex",
-  },
-  {
-    name: "Faubourg Tower",
-    coordinate: {
-      latitude: 45.494666,
-      longitude: -73.577603,
-    },
-    address: "1250 Guy Street, Montreal, QC H3H 2L4",
-    fullBuildingName: "Faubourg Tower",
-  },
-  {
-    name: "FG Building",
-    coordinate: {
-      latitude: 45.494381,
-      longitude: -73.578425,
-    },
-    address: "1610 Ste-Catherine, Montreal, QC H3H 2S2",
-    fullBuildingName: "Faubourg Building",
-  },
-  {
-    name: "Grey Nuns Annex",
-    coordinate: {
-      latitude: 45.494123,
-      longitude: -73.57787,
-    },
-    address: "1211-1215 St-Mathieu, Montreal, QC H3H 2P9",
-    fullBuildingName: "Grey Nuns Annex",
-  },
-  {
-    name: "Guy-De Maisonneuve Building",
-    coordinate: {
-      latitude: 45.495983,
-      longitude: -73.578824,
-    },
-    address: "1550 De Maisonneuve W, Montreal, QC H3G 1M8",
-    fullBuildingName: "Guy-De Maisonneuve Building",
-  },
-  {
-    name: "Grey Nuns Building - Main Entrance",
-    coordinate: {
-      latitude: 45.4935,
-      longitude: -73.576873,
-    },
-    address: "1190 Guy Street, Montreal, QC H3H 2L4",
-    fullBuildingName: "Grey Nuns Building - Main Entrance",
-  },
-  {
-    name: "GS Building",
-    coordinate: {
-      latitude: 45.496673,
-      longitude: -73.581409,
-    },
-    address: "1538 Sherbrooke W, Montreal, QC H3G 1H6",
-    fullBuildingName: "Grey Sisters Building",
-  },
-  {
-    name: "H Building",
-    coordinate: {
-      latitude: 45.497092,
-      longitude: -73.5788,
-    },
-    address: "1455 De Maisonneuve W, Montreal, QC H3G 1M8",
-    fullBuildingName: "Henry F. Hall Building",
-  },
-  {
-    name: "K Building",
-    coordinate: {
-      latitude: 45.497803,
-      longitude: -73.579357,
-    },
-    address: "2150 Bishop Street, Montreal, QC H3G 2E9",
-    fullBuildingName: "K Annex",
-  },
-  {
-    name: "LC Building",
-    coordinate: {
-      latitude: 45.496782,
-      longitude: -73.577358,
-    },
-    address: "1426 Bishop Street, Montreal, QC H3G 2E8",
-    fullBuildingName: "Library and Computing Building",
-  },
-  {
-    name: "LD Building",
-    coordinate: {
-      latitude: 45.496697,
-      longitude: -73.577312,
-    },
-    address: "1424 Bishop Street, Montreal, QC H3G 2E8",
-    fullBuildingName: "Library and Documentation Building",
-  },
-  {
-    name: "Learning Square",
-    coordinate: {
-      latitude: 45.4963149,
-      longitude: -73.579472,
-    },
-    address: "1400 De Maisonneuve W, Montreal, QC H3G 1M8",
-    fullBuildingName: "Learning Square",
-  },
-  {
-    name: "M Annex",
-    coordinate: {
-      latitude: 45.497368,
-      longitude: -73.579777,
-    },
-    address: "2135 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Mackay Annex",
-  },
-  {
-    name: "JMSB",
-    coordinate: {
-      latitude: 45.495304,
-      longitude: -73.579044,
-    },
-    address: "1450 Guy Street, Montreal, QC H3H 0A1",
-    fullBuildingName: "John Molson School of Business Building",
-  },
-  {
-    name: "MI Annex",
-    coordinate: {
-      latitude: 45.490889,
-      longitude: -73.582412,
-    },
-    address: "2130 Bishop Street, Montreal, QC H3G 2E9",
-    fullBuildingName: "Montreal International Annex",
-  },
-  {
-    name: "MK Annex",
-    coordinate: {
-      latitude: 45.496606,
-      longitude: -73.579025,
-    },
-    address: "2000-20002 Mackay St, Montreal, QC H3G 2J1",
-    fullBuildingName: "Mackay-Kilgour Annex",
-  },
-  {
-    name: "MM Annex",
-    coordinate: {
-      latitude: 45.494665,
-      longitude: -73.576315,
-    },
-    address: "1209 Guy Street, Montreal, QC H3H 2L4",
-    fullBuildingName: "Montreal Metro Annex",
-  },
-  {
-    name: "MN Annex",
-    coordinate: {
-      latitude: 45.494568,
-      longitude: -73.579553,
-    },
-    address: "1205-1207 Guy Street, Montreal, QC H3H 2L4",
-    fullBuildingName: "Montreal North Annex",
-  },
-  {
-    name: "MT Building",
-    coordinate: {
-      latitude: 45.494442,
-      longitude: -73.576108,
-    },
-    address: "1195 Guy Street, Montreal, QC H3H 2L4",
-    fullBuildingName: "Montefiore Building",
-  },
-  {
-    name: "MU Annex",
-    coordinate: {
-      latitude: 45.497963,
-      longitude: -73.579506,
-    },
-    address: "2170 Bishop Street, Montreal, QC H3G 2E9",
-    fullBuildingName: "Music Annex",
-  },
-  {
-    name: "P Annex",
-    coordinate: {
-      latitude: 45.496661,
-      longitude: -73.579165,
-    },
-    address: "2020 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Psychology Annex",
-  },
-  {
-    name: "PR Annex",
-    coordinate: {
-      latitude: 45.49697,
-      longitude: -73.579851,
-    },
-    address: "2100 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Public Relations Annex",
-  },
-  {
-    name: "Q Annex",
-    coordinate: {
-      latitude: 45.496648,
-      longitude: -73.579094,
-    },
-    address: "2010 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Quebec Annex",
-  },
-  {
-    name: "R Annex",
-    coordinate: {
-      latitude: 45.496826,
-      longitude: -73.579389,
-    },
-    address: "2050 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Research Annex",
-  },
-  {
-    name: "RR Annex",
-    coordinate: {
-      latitude: 45.497407,
-      longitude: -73.579848,
-    },
-    address: "2040 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Renaissance Research Annex",
-  },
-  {
-    name: "S Annex",
-    coordinate: {
-      latitude: 45.497423,
-      longitude: -73.579851,
-    },
-    address: "2145 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Science Annex",
-  },
-  {
-    name: "Samuel Bronfman Building, SB Building",
-    coordinate: {
-      latitude: 45.4966,
-      longitude: -73.58609,
-    },
-    address: "1590 Doctor Penfield, Montreal, QC H3G 1C5",
-    fullBuildingName: "Samuel Bronfman Building",
-  },
-  {
-    name: "Toronto-Dominion Building",
-    coordinate: {
-      latitude: 45.494993,
-      longitude: -73.578129,
-    },
-    address: "1410 Guy Street, Montreal, QC H3H 2L7",
-    fullBuildingName: "Toronto-Dominion Building",
-  },
-  {
-    name: "TU Tunnel",
-    coordinate: {
-      latitude: 45.49648,
-      longitude: -73.578918,
-    },
-    address: "1550 De Maisonneuve W, Montreal, QC H3G 1M8",
-    fullBuildingName: "Tunnel Underground",
-  },
-  {
-    name: "V Annex",
-    coordinate: {
-      latitude: 45.497101,
-      longitude: -73.579907,
-    },
-    address: "2110 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Visual Arts Annex",
-  },
-  {
-    name: "Visual Art Building",
-    coordinate: {
-      latitude: 45.495543,
-      longitude: -73.573795,
-    },
-    address: "1395 Rene Levesque W, Montreal, QC H3G 2M5",
-    fullBuildingName: "Visual Arts Building",
-  },
-  {
-    name: "X Building",
-    coordinate: {
-      latitude: 45.49694,
-      longitude: -73.579593,
-    },
-    address: "2080 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "X Annex",
-  },
-  {
-    name: "Z Building",
-    coordinate: {
-      latitude: 45.496981,
-      longitude: -73.579705,
-    },
-    address: "2090 Mackay Street, Montreal, QC H3G 2J1",
-    fullBuildingName: "Z Annex",
-  },
-
-  // Loyola Campus buildings latitude and longitude
-  {
-    name: "Administration Building",
-    coordinate: {
-      latitude: 45.457984,
-      longitude: -73.639834,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Administration Building",
-  },
-  {
-    name: "BB Annex",
-    coordinate: {
-      latitude: 45.459754,
-      longitude: -73.639176,
-    },
-    address: "3502 Bermore Avenue, Montreal, QC H4B 2A3",
-    fullBuildingName: "Bermore Annex",
-  },
-  {
-    name: "BH Annex",
-    coordinate: {
-      latitude: 45.459819,
-      longitude: -73.639152,
-    },
-    address: "3500 Bermore Avenue, Montreal, QC H4B 2A3",
-    fullBuildingName: "Bermore Hall Annex",
-  },
-  {
-    name: "Central Building",
-    coordinate: {
-      latitude: 45.458204,
-      longitude: -73.6403,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Central Building",
-  },
-  {
-    name: "Communication Studies and Journalism Building",
-    coordinate: {
-      latitude: 45.457478,
-      longitude: -73.640354,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Communication Studies and Journalism Building",
-  },
-  {
-    name: "Stinger Dome",
-    coordinate: {
-      latitude: 45.457656,
-      longitude: -73.636098,
-    },
-    address: "7200 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Stinger Dome",
-  },
-  {
-    name: "F.C. Building",
-    coordinate: {
-      latitude: 45.458493,
-      longitude: -73.639287,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "F.C. Smith Building",
-  },
-  {
-    name: "Centre for Structural and Functional Genomics",
-    coordinate: {
-      latitude: 45.457017,
-      longitude: -73.640432,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Centre for Structural and Functional Genomics",
-  },
-  {
-    name: "Hingston Hall",
-    coordinate: {
-      latitude: 45.459356,
-      longitude: -73.64127,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Hingston Hall - Wing HA",
-  },
-  {
-    name: "Hingston Hall",
-    coordinate: {
-      latitude: 45.459162,
-      longitude: -73.641962,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Hingston Hall - Wing HB",
-  },
-  {
-    name: "Hingston Hall",
-    coordinate: {
-      latitude: 45.458504,
-      longitude: -73.643225,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Hingston Hall - Wing HC",
-  },
-  {
-    name: "Applied Science Hub, HU Building",
-    coordinate: {
-      latitude: 45.458513,
-      longitude: -73.641921,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Applied Science Hub",
-  },
-  {
-    name: "Jesuit Residence",
-    coordinate: {
-      latitude: 45.458504,
-      longitude: -73.643225,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Jesuit Residence",
-  },
-  {
-    name: "PB Building",
-    coordinate: {
-      latitude: 45.456534,
-      longitude: -73.638106,
-    },
-    address: "7200 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Physical Sciences Building",
-  },
-  {
-    name: "Oscar Peterson Concert Hall, PT Building",
-    coordinate: {
-      latitude: 45.459308,
-      longitude: -73.638941,
-    },
-    address: "7141 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Oscar Peterson Concert Hall",
-  },
-
-  {
-    name: "Perform Center",
-    coordinate: {
-      latitude: 45.457088,
-      longitude: -73.637683,
-    },
-    address: "7200 Sherbrooke West, Montreal, QC H4B 1R6",
-    fullBuildingName: "Perform Center",
-  },
-  {
-    name: "Physical Service",
-    coordinate: {
-      latitude: 45.459636,
-      longitude: -73.639758,
-    },
-    address: "7141 Sherbrooke West",
-    fullBuildingName: "Physical Services Building",
-  },
-  {
-    name: "Psychology Building",
-    coordinate: {
-      latitude: 45.458938,
-      longitude: -73.640467,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Psychology Building",
-  },
-  {
-    name: "Recreational and Athletics",
-    coordinate: {
-      latitude: 45.456677,
-      longitude: -73.637523,
-    },
-    address: "7200 Sherbrooke W",
-    fullBuildingName: "Recreation and Athletics Building",
-  },
-  {
-    name: "Loyola Jesuit Hall and Conference Center",
-    coordinate: {
-      latitude: 45.458551,
-      longitude: -73.641016,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Loyola Jesuit Hall and Conference Center",
-  },
-  {
-    name: "Student Center",
-    coordinate: {
-      latitude: 45.459137,
-      longitude: -73.639203,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Student Center",
-  },
-  {
-    name: "SH Building",
-    coordinate: {
-      latitude: 45.459298,
-      longitude: -73.642478,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Solar House",
-  },
-  {
-    name: "TA Building",
-    coordinate: {
-      latitude: 45.460051,
-      longitude: -73.640842,
-    },
-    address: "7079 Terrebonne",
-    fullBuildingName: "Terrebonne Building",
-  },
-  {
-    name: "TB Annex",
-    coordinate: {
-      latitude: 45.459992,
-      longitude: -73.640897,
-    },
-    address: "7075 Terrebonne",
-    fullBuildingName: "Terrebonne Annex",
-  },
-  {
-    name: "Richard J. Renaud Science Complex",
-    coordinate: {
-      latitude: 45.457567,
-      longitude: -73.641739,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Richard J. Renaud Science Complex",
-  },
-  {
-    name: "Vanier Library",
-    coordinate: {
-      latitude: 45.459026,
-      longitude: -73.638606,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Vanier Library",
-  },
-  {
-    name: "Vanier Extension",
-    coordinate: {
-      latitude: 45.459026,
-      longitude: -73.638606,
-    },
-    address: "7141 Sherbrooke W",
-    fullBuildingName: "Vanier Library Extension",
-  },
-];
-
-// Default center coordinates for each campus
-export const CampusCoordinates = {
-  SGW: {
-    latitude: 45.4973,
-    longitude: -73.5789,
-    latitudeDelta: 0.01,
-    longitudeDelta: 0.01,
-  },
-  Loyola: {
-    latitude: 45.458256,
-    longitude: -73.640472,
-    latitudeDelta: 0.01,
-    longitudeDelta: 0.01,
-  },
-};
