+++ ConcordiaMaps/constants/coordinates/CampusCoordinates.js
@@ -1,0 +1,14 @@
+export const CampusCoordinates = {
+    SGW: {
+      latitude: 45.4973,
+      longitude: -73.5789,
+      latitudeDelta: 0.01,
+      longitudeDelta: 0.01,
+    },
+    Loyola: {
+      latitude: 45.458256,
+      longitude: -73.640472,
+      latitudeDelta: 0.01,
+      longitudeDelta: 0.01,
+    },
+  };
