+++ ConcordiaMaps/screen/HomeScreen.js
@@ -8,7 +8,7 @@ import { LocationContext } from "../contexts/LocationContext";
 import Footer from "../components/Footer";
 import styles from "../styles";
 const customMarkerImage = require("../assets/PinLogo.png");
-import { Building } from "../components/MapMarkers";
+import { Building } from "../constants/Building";
 import { ModalContext } from "../App";
 import BuildingColoring from "../components/buildingColoring";
 import Legend from "../components/Legend";
@@ -21,6 +21,7 @@ import {
 } from "../components/AsyncPersistence";
 import convertToCoordinates from "../components/convertToCoordinates";
 import PropTypes from "prop-types";
+import MapMarkers from "../components/MapMarkers";
 
 function HomeScreen({ asyncKey = "Campus" }) {
   const loyolaPostalCode = process.env.EXPO_PUBLIC_LOYOLA_POSTAL_CODE;
@@ -175,7 +176,7 @@ function HomeScreen({ asyncKey = "Campus" }) {
             watchUserLocation={true}
             onRegionChangeComplete={(region) => setMapRegion(region)}
           >
-            {Building.map((building) => (
+            {/* {Building.map((building) => (
               <Marker
                 key={`${building.name}-${building.coordinate.latitude}-${building.coordinate.longitude}`}
                 testID={`marker-${building.name?.toLowerCase().replace(/\s+/g, "-") || building.id}`}
@@ -190,7 +191,8 @@ function HomeScreen({ asyncKey = "Campus" }) {
                   style={styles.customMarkerImage}
                 />
               </Marker>
-            ))}
+            ))} */}
+            <MapMarkers/>
             <BuildingColoring />
             {selectedLocation && (
               <Marker
